refactor: move type movie interface to types folder
