+++ src/types/interfaces.ts
@@ -1,0 +1,7 @@
+export interface Movie {
+  id: string;
+  name: string;
+  genres: string[];
+  year: number;
+  poster: string;
+}
